Add `@aws-lite/s3` `GetObject` method Move common `@aws-lite/s3` code into a lib file
